revert split across files
