+++ mesa/space.py
@@ -1,13 +1,23 @@
-# -*- coding: utf-8 -*-
 """
 Mesa Space Module
 =================
 
-Objects used to add a spatial component to a model.
+Classes used to add a spatial component to a model.
+
+Grids
+-----
 
 Grid: base grid, a simple list-of-lists.
 SingleGrid: grid which strictly enforces one object per cell.
 MultiGrid: extension to Grid where each cell is a set of objects.
+
+HexGrid: Extends Grid to handle hexagonal neighbors.
+
+Other Spaces
+------------
+
+ContinuousSpace: Continuous space where each agent can have an arbitrary position.
+NetworkGrid: A Network of nodes based on networkx
 
 """
 # Instruction for PyLint to suppress variable name errors, since we have a
@@ -15,63 +25,72 @@ MultiGrid: extension to Grid where each cell is a set of objects.
 # pylint: disable=invalid-name
 
 import itertools
+import warnings
+from typing import (
+    Any,
+    Callable,
+    Dict,
+    Iterable,
+    Iterator,
+    List,
+    Optional,
+    Set,
+    Tuple,
+    TypeVar,
+    Union,
+    cast,
+)
 
 import numpy as np
 
-from typing import Iterable, Iterator, List, Optional, Set, Tuple, Union
-from .agent import Agent
+from mesa.agent import Agent
+
 Coordinate = Tuple[int, int]
-GridContent = Union[Optional[Agent], Set[Agent]]
+GridContent = List[Agent]
+
+F = TypeVar("F", bound=Callable[..., Any])
 
 
-def accept_tuple_argument(wrapped_function):
+def accept_tuple_argument(wrapped_function: F) -> F:
     """ Decorator to allow grid methods that take a list of (x, y) coord tuples
     to also handle a single position, by automatically wrapping tuple in
     single-item list rather than forcing user to do it.
 
     """
 
-    def wrapper(*args):
+    def wrapper(*args: Any) -> Any:
         if isinstance(args[1], tuple) and len(args[1]) == 2:
             return wrapped_function(args[0], [args[1]])
-        else:
-            return wrapped_function(*args)
+        return wrapped_function(*args)
 
-    return wrapper
+    return cast(F, wrapper)
 
 
-class Grid:
-    """ Base class for a square grid.
+class MultiGrid:
+    """ Base class for a rectangular grid with square cells.
 
     Grid cells are indexed by [x][y], where [0][0] is assumed to be the
     bottom-left and [width-1][height-1] is the top-right. If a grid is
     toroidal, the top and bottom, and left and right, edges wrap to each other
+    Each position of the grid is referred to by a (x, y) coordinate tuple.
+    You may access the content of a single cell by calling Grid[x, y].
 
     Properties:
         width, height: The grid's width and height.
         torus: Boolean which determines whether to treat the grid as a torus.
-        grid: Internal list-of-lists which holds the grid cells themselves.
+        empties: List of currently empty cells.
 
     Methods:
+        position_agent: Preferred function to initially place agents.
+        place_agent: Positions an agent on the grid, and set its pos variable.
+        move_agent: Moves an agent from its current to a new position.
+        remove_agent: Remove an agent from the grid.
         get_neighbors: Returns the objects surrounding a given cell.
         get_neighborhood: Returns the cells surrounding a given cell.
-        get_cell_list_contents: Returns the contents of a list of cells
-            ((x,y) tuples)
-        neighbor_iter: Iterates over position neightbors.
-        coord_iter: Returns coordinates as well as cell contents.
-        place_agent: Positions an agent on the grid, and set its pos variable.
-        move_agent: Moves an agent from its current position to a new position.
-        iter_neighborhood: Returns an iterator over cell coordinates that are
-        in the neighborhood of a certain point.
-        torus_adj: Converts coordinate, handles torus looping.
-        out_of_bounds: Determines whether position is off the grid, returns
-        the out of bounds coordinate.
-        iter_cell_list_contents: Returns an iterator of the contents of the
-        cells identified in cell_list.
-        get_cell_list_contents: Returns a list of the contents of the cells
-        identified in cell_list.
-        remove_agent: Removes an agent from the grid.
-        is_cell_empty: Returns a bool of the contents of a cell.
+        get_contents: Returns the contents of a list of cells.
+        coord_iter: Returns cell contens and coordinates of all cells.
+        torus_adj: Converts coordinates, handles torus looping.
+        out_of_bounds: Determines whether position is off the grid.
 
     """
 
@@ -87,99 +106,103 @@ class Grid:
         self.width = width
         self.torus = torus
 
-        self.grid = []  # type: List[List[GridContent]]
+        self._grid = []  # type: List[List[GridContent]]
 
-        for x in range(self.width):
+        for _ in range(self.width):
             col = []  # type: List[GridContent]
-            for y in range(self.height):
-                col.append(self.default_val())
-            self.grid.append(col)
+            for _ in range(self.height):
+                col.append([])
+            self._grid.append(col)
 
         # Add all cells to the empties list.
-        self.empties = set(itertools.product(
-            *(range(self.width), range(self.height))))
+        self._empties = set(itertools.product(range(self.width), range(self.height)))
+        self._all_cells = frozenset(self._empties)
 
-    @staticmethod
-    def default_val() -> None:
-        """ Default value for new cell elements. """
-        return None
+        # Neighborhood Cache
+        self._neighborhood_cache: Dict[Any, List[Coordinate]] = dict()
 
-    def __getitem__(self, index: int) -> List[GridContent]:
-        return self.grid[index]
+    def __getitem__(self, pos: Coordinate) -> GridContent:
+        """Access contents of a given position."""
+        if isinstance(pos, int):
+            warnings.warn(
+                """Accesing the grid via `grid[x][y]` is depreciated.
+                Use `grid[x, y]` instead."""
+            )
+            return self._grid[pos]
+        x, y = pos
+        return self._grid[x][y]
+
+    def __setitem__(self, pos: Coordinate, agent: Agent) -> None:
+        """Add agents to a position."""
+        x, y = pos
+        self._grid[x][y].append(agent)
 
     def __iter__(self) -> Iterator[GridContent]:
-        """
-        create an iterator that chains the
-        rows of grid together as if one list:
-        """
-        return itertools.chain(*self.grid)
+        """Iterate over all cells in the grid."""
+        return itertools.chain.from_iterable(self._grid)
 
     def coord_iter(self) -> Iterator[Tuple[GridContent, int, int]]:
-        """ An iterator that returns coordinates as well as cell contents. """
+        """Iterate over all cell contents and coordinates. """
         for row in range(self.width):
             for col in range(self.height):
-                yield self.grid[row][col], row, col  # agent, x, y
+                yield self[row, col], row, col  # agent, x, y
 
-    def neighbor_iter(self, pos: Coordinate, moore: bool = True) -> Iterator[GridContent]:
-        """ Iterate over position neighbors.
+    def torus_adj(self, pos: Coordinate) -> Coordinate:
+        """Convert coordinates, handling torus looping."""
+        if not self.out_of_bounds(pos):
+            return pos
+        if not self.torus:
+            raise Exception("Point out of bounds, and space non-toroidal.")
+        return pos[0] % self.width, pos[1] % self.height
 
-        Args:
-            pos: (x,y) coords tuple for the position to get the neighbors of.
-            moore: Boolean for whether to use Moore neighborhood (including
-                   diagonals) or Von Neumann (only up/down/left/right).
+    def out_of_bounds(self, pos: Coordinate) -> bool:
+        """Determines whether position is off the grid."""
+        return tuple(pos) not in self._all_cells
 
-        """
-        neighborhood = self.iter_neighborhood(pos, moore=moore)
-        return self.iter_cell_list_contents(neighborhood)
-
-    def iter_neighborhood(self, pos: Coordinate, moore: bool,
-                          include_center: bool = False, radius: int = 1) -> Iterator[Coordinate]:
-        """ Return an iterator over cell coordinates that are in the
-        neighborhood of a certain point.
-
-        Args:
-            pos: Coordinate tuple for the neighborhood to get.
-            moore: If True, return Moore neighborhood
-                        (including diagonals)
-                   If False, return Von Neumann neighborhood
-                        (exclude diagonals)
-            include_center: If True, return the (x, y) cell as well.
-                            Otherwise, return surrounding cells only.
-            radius: radius, in cells, of neighborhood to get.
-
-        Returns:
-            A list of coordinate tuples representing the neighborhood. For
-            example with radius 1, it will return list with number of elements
-            equals at most 9 (8) if Moore, 5 (4) if Von Neumann (if not
-            including the center).
-
-        """
+    def place_agent(self, agent: Agent, pos: Coordinate) -> Agent:
+        """Position an agent on the grid, and set its pos variable."""
         x, y = pos
-        coordinates = set()  # type: Set[Coordinate]
-        for dy in range(-radius, radius + 1):
-            for dx in range(-radius, radius + 1):
-                if dx == 0 and dy == 0 and not include_center:
-                    continue
-                # Skip coordinates that are outside manhattan distance
-                if not moore and abs(dx) + abs(dy) > radius:
-                    continue
-                # Skip if not a torus and new coords out of bounds.
-                if not self.torus and (not (0 <= dx + x < self.width) or not (0 <= dy + y < self.height)):
-                    continue
+        self._grid[x][y].append(agent)
+        self._empties.discard(pos)
+        setattr(agent, "pos", pos)
+        return agent
 
-                px, py = self.torus_adj((x + dx, y + dy))
+    def remove_agent(self, agent: Agent) -> Agent:
+        """Remove the agent from the grid and set its pos variable to None."""
+        x, y = getattr(agent, "pos")
+        content = self._grid[x][y]
+        content.remove(agent)
+        if not content:
+            self._empties.add((x, y))
+        setattr(agent, "pos", None)
+        return agent
 
-                # Skip if new coords out of bounds.
-                if self.out_of_bounds((px, py)):
-                    continue
+    def move_agent(self, agent: Agent, pos: Coordinate) -> Agent:
+        """
+        Move an agent from its current position to a new position.
 
-                coords = (px, py)
-                if coords not in coordinates:
-                    coordinates.add(coords)
-                    yield coords
+        Args:
+            agent: Agent to move. Must have a valid pos attribute.
+            pos: Tuple of new position to move the agent to.
 
-    def get_neighborhood(self, pos: Coordinate, moore: bool,
-                         include_center: bool = False, radius: int = 1) -> List[Coordinate]:
+        """
+        pos = self.torus_adj(pos)
+        self.remove_agent(agent)
+        self.place_agent(agent, pos)
+        return agent
+
+    @accept_tuple_argument
+    def get_contents(self, cell_list: Iterable[Coordinate]) -> List[GridContent]:
+        """Return a list of the cell contents for a given cell list."""
+        return [self[pos] for pos in cell_list if not self.is_cell_empty(pos)]
+
+    def get_neighborhood(
+        self,
+        pos: Coordinate,
+        moore: bool,
+        include_center: bool = False,
+        radius: int = 1,
+    ) -> List[Coordinate]:
         """ Return a list of cells that are in the neighborhood of a
         certain point.
 
@@ -199,35 +222,40 @@ class Grid:
             if not including the center).
 
         """
-        return list(self.iter_neighborhood(pos, moore, include_center, radius))
+        cache_key = (pos, moore, include_center, radius)
+        neighborhood = self._neighborhood_cache.get(cache_key, None)
+        if neighborhood is None:
+            x, y = pos
+            coordinates = set()  # type: Set[Coordinate]
+            for dy in range(-radius, radius + 1):
+                for dx in range(-radius, radius + 1):
+                    if dx == 0 and dy == 0 and not include_center:
+                        continue
+                    # Skip coordinates that are outside manhattan distance
+                    if not moore and abs(dx) + abs(dy) > radius:
+                        continue
+                    # Skip if not a torus and new coords out of bounds.
+                    coord = (x + dx, y + dy)
 
-    def iter_neighbors(self, pos: Coordinate, moore: bool,
-                       include_center: bool = False, radius: int = 1) -> Iterator[GridContent]:
-        """ Return an iterator over neighbors to a certain point.
+                    if self.out_of_bounds(coord):
+                        if not self.torus:
+                            continue
+                        coord = self.torus_adj(coord)
 
-        Args:
-            pos: Coordinates for the neighborhood to get.
-            moore: If True, return Moore neighborhood
-                    (including diagonals)
-                   If False, return Von Neumann neighborhood
-                     (exclude diagonals)
-            include_center: If True, return the (x, y) cell as well.
-                            Otherwise,
-                            return surrounding cells only.
-            radius: radius, in cells, of neighborhood to get.
+                    if coord not in coordinates:
+                        coordinates.add(coord)
 
-        Returns:
-            An iterator of non-None objects in the given neighborhood;
-            at most 9 if Moore, 5 if Von-Neumann
-            (8 and 4 if not including the center).
+            neighborhood = sorted(coordinates)
+            self._neighborhood_cache[cache_key] = neighborhood
+        return neighborhood
 
-        """
-        neighborhood = self.iter_neighborhood(
-            pos, moore, include_center, radius)
-        return self.iter_cell_list_contents(neighborhood)
-
-    def get_neighbors(self, pos: Coordinate, moore: bool,
-                      include_center: bool = False, radius: int = 1) -> List[Coordinate]:
+    def get_neighbors(
+        self,
+        pos: Coordinate,
+        moore: bool = True,
+        include_center: bool = False,
+        radius: int = 1,
+    ) -> List[Agent]:
         """ Return a list of neighbors to a certain point.
 
         Args:
@@ -247,130 +275,95 @@ class Grid:
             (8 and 4 if not including the center).
 
         """
-        return list(self.iter_neighbors(
-            pos, moore, include_center, radius))
+        neighborhood = self.get_neighborhood(pos, moore, include_center, radius)
+        neighbors = self.get_contents(neighborhood)
+        return list(itertools.chain.from_iterable(neighbors))
 
-    def torus_adj(self, pos: Coordinate) -> Coordinate:
-        """ Convert coordinate, handling torus looping. """
-        if not self.out_of_bounds(pos):
-            return pos
-        elif not self.torus:
-            raise Exception("Point out of bounds, and space non-toroidal.")
-        else:
-            x, y = pos[0] % self.width, pos[1] % self.height
-        return x, y
+    def neighbor_iter(self, pos: Coordinate, moore: bool = True) -> Iterator[Agent]:
+        """Depreciated."""
+        yield from self.get_neighbors(pos, moore=moore)
 
-    def out_of_bounds(self, pos: Coordinate) -> bool:
-        """
-        Determines whether position is off the grid, returns the out of
-        bounds coordinate.
-        """
-        x, y = pos
-        return x < 0 or x >= self.width or y < 0 or y >= self.height
+    def iter_neighborhood(
+        self,
+        pos: Coordinate,
+        moore: bool,
+        include_center: bool = False,
+        radius: int = 1,
+    ) -> Iterator[Coordinate]:
+        """Depreciated."""
+        yield from self.get_neighborhood(pos, moore, include_center, radius)
 
-    @accept_tuple_argument
-    def iter_cell_list_contents(self, cell_list: Iterable[Coordinate]) -> Iterator[GridContent]:
-        """
-        Args:
-            cell_list: Array-like of (x, y) tuples, or single tuple.
-
-        Returns:
-            An iterator of the contents of the cells identified in cell_list
-
-        """
-        return (
-            self[x][y] for x, y in cell_list if not self.is_cell_empty((x, y)))
+    def iter_neighbors(
+        self,
+        pos: Coordinate,
+        moore: bool,
+        include_center: bool = False,
+        radius: int = 1,
+    ) -> Iterator[GridContent]:
+        """Depreciated."""
+        neighborhood = self.get_neighborhood(pos, moore, include_center, radius)
+        yield from self.get_contents(neighborhood)
 
     @accept_tuple_argument
-    def get_cell_list_contents(self, cell_list: Iterable[Coordinate]) -> List[GridContent]:
-        """
-        Args:
-            cell_list: Array-like of (x, y) tuples, or single tuple.
+    def iter_cell_list_contents(
+        self, cell_list: Iterable[Coordinate]
+    ) -> Iterator[GridContent]:
+        """Depreciated."""
+        yield from self.get_contents(cell_list)
 
-        Returns:
-            A list of the contents of the cells identified in cell_list
-
-        """
-        return list(self.iter_cell_list_contents(cell_list))
-
-    def move_agent(self, agent: Agent, pos: Coordinate) -> None:
-        """
-        Move an agent from its current position to a new position.
-
-        Args:
-            agent: Agent object to move. Assumed to have its current location
-                   stored in a 'pos' tuple.
-            pos: Tuple of new position to move the agent to.
-
-        """
-        pos = self.torus_adj(pos)
-        self._remove_agent(agent.pos, agent)
-        self._place_agent(pos, agent)
-        agent.pos = pos
-
-    def place_agent(self, agent: Agent, pos: Coordinate) -> None:
-        """ Position an agent on the grid, and set its pos variable. """
-        self._place_agent(pos, agent)
-        agent.pos = pos
-
-    def _place_agent(self, pos: Coordinate, agent: Agent) -> None:
-        """ Place the agent at the correct location. """
-        x, y = pos
-        self.grid[x][y] = agent
-        self.empties.discard(pos)
-
-    def remove_agent(self, agent: Agent) -> None:
-        """ Remove the agent from the grid and set its pos variable to None. """
-        pos = agent.pos
-        self._remove_agent(pos, agent)
-        agent.pos = None
-
-    def _remove_agent(self, pos: Coordinate, agent: Agent) -> None:
-        """ Remove the agent from the given location. """
-        x, y = pos
-        self.grid[x][y] = None
-        self.empties.add(pos)
+    @accept_tuple_argument
+    def get_cell_list_contents(self, cell_list: Iterable[Coordinate]) -> List[Agent]:
+        """Depreciated"""
+        return list(itertools.chain(*self.get_contents(cell_list)))
 
     def is_cell_empty(self, pos: Coordinate) -> bool:
         """ Returns a bool of the contents of a cell. """
-        x, y = pos
-        return self.grid[x][y] == self.default_val()
+        return pos in self._empties
 
     def move_to_empty(self, agent: Agent) -> None:
         """ Moves agent to a random empty cell, vacating agent's old cell. """
-        pos = agent.pos
-        if len(self.empties) == 0:
+        if len(self._empties) == 0:
             raise Exception("ERROR: No empty cells")
-        new_pos = agent.random.choice(sorted(self.empties))
-        self._place_agent(new_pos, agent)
-        agent.pos = new_pos
-        self._remove_agent(pos, agent)
+        new_pos = agent.random.choice(self.empties)
+        self.move_agent(agent, new_pos)
 
     def find_empty(self) -> Optional[Coordinate]:
         """ Pick a random empty cell. """
         from warnings import warn
         import random
 
-        warn(("`find_empty` is being phased out since it uses the global "
-              "`random` instead of the model-level random-number generator. "
-              "Consider replacing it with having a model or agent object "
-              "explicitly pick one of the grid's list of empty cells."),
-             DeprecationWarning)
+        warn(
+            (
+                "`find_empty` is being phased out since it uses the global "
+                "`random` instead of the model-level random-number generator. "
+                "Consider replacing it with having a model or agent object "
+                "explicitly pick one of the grid's list of empty cells."
+            ),
+            DeprecationWarning,
+        )
 
         if self.exists_empty_cells():
-            pos = random.choice(sorted(self.empties))
-            return pos
-        else:
-            return None
+            return random.choice(self.empties)
+        return None
+
+    @property
+    def empties(self) -> List[Coordinate]:
+        return sorted(self._empties)
+
+    @property
+    def all_cells(self) -> List[Coordinate]:
+        return sorted(self._all_cells)
 
     def exists_empty_cells(self) -> bool:
-        """ Return True if any cells empty else False. """
-        return len(self.empties) > 0
+        """Depreciated.
+
+        Test with `if grid.empties`
+        """
+        return len(self._empties) > 0
 
 
-class SingleGrid(Grid):
+class SingleGrid(MultiGrid):
     """ Grid where each cell contains exactly at most one object. """
-    empties = set()  # type: Set[Coordinate]
 
     def __init__(self, width: int, height: int, torus: bool) -> None:
         """ Create a new single-item grid.
@@ -382,7 +375,20 @@ class SingleGrid(Grid):
         """
         super().__init__(width, height, torus)
 
-    def position_agent(self, agent, x="random", y="random"):
+    def __getitem__(self, pos: Coordinate) -> Optional[Agent]:
+        if isinstance(pos, int):
+            warnings.warn("depreciated")
+            return [content[0] for content in self._grid[pos] if content]
+        x, y = pos
+        content = self._grid[x][y]
+        return content[0] if content else None
+
+    def get_contents(self, cell_list: Iterable[Coordinate]) -> List[Agent]:
+        return super().get_contents(cell_list)
+
+    def position_agent(
+        self, agent: Agent, x: Union[str, int] = "random", y: Union[str, int] = "random"
+    ) -> None:
         """ Position an agent on the grid.
         This is used when first placing agents! Use 'move_to_empty()'
         when you want agents to jump to an empty cell.
@@ -392,23 +398,57 @@ class SingleGrid(Grid):
         Ensure this random position is not occupied (in Grid).
 
         """
+        # TODO: Allow to use only one random value
         if x == "random" or y == "random":
-            if len(self.empties) == 0:
+            if len(self._empties) == 0:
                 raise Exception("ERROR: Grid full")
-            coords = agent.random.choice(sorted(self.empties))
+            coords: Tuple[int, int] = agent.random.choice(self.empties)
         else:
-            coords = (x, y)
+            coords = (int(x), int(y))
         agent.pos = coords
-        self._place_agent(coords, agent)
+        self.place_agent(agent, coords)
 
-    def _place_agent(self, pos: Coordinate, agent: Agent) -> None:
-        if self.is_cell_empty(pos):
-            super()._place_agent(pos, agent)
-        else:
+    def place_agent(self, agent: Agent, pos: Coordinate) -> Agent:
+        if not self.is_cell_empty(pos):
             raise Exception("Cell not empty")
+        return super().place_agent(agent, pos)
+
+    def get_neighbors(
+        self,
+        pos: Coordinate,
+        moore: bool = True,
+        include_center: bool = False,
+        radius: int = 1,
+    ) -> List[Agent]:
+        """ Return a list of neighbors to a certain point.
+
+        Args:
+            pos: Coordinate tuple for the neighborhood to get.
+            moore: If True, return Moore neighborhood
+                    (including diagonals)
+                   If False, return Von Neumann neighborhood
+                     (exclude diagonals)
+            include_center: If True, return the (x, y) cell as well.
+                            Otherwise,
+                            return surrounding cells only.
+            radius: radius, in cells, of neighborhood to get.
+
+        Returns:
+            A list of non-None objects in the given neighborhood;
+            at most 9 if Moore, 5 if Von-Neumann
+            (8 and 4 if not including the center).
+
+        """
+        neighborhood = self.get_neighborhood(pos, moore, include_center, radius)
+        return self.get_contents(neighborhood)
+        # return list(itertools.chain.from_iterable(neighbors))
+
+    @accept_tuple_argument
+    def get_cell_list_contents(self, cell_list: Iterable[Coordinate]) -> List[Agent]:
+        return self.get_contents(cell_list)
 
 
-class MultiGrid(Grid):
+class Grid(SingleGrid):
     """ Grid where each cell can contain more than one object.
 
     Grid cells are indexed by [x][y], where [0][0] is assumed to be at
@@ -428,36 +468,12 @@ class MultiGrid(Grid):
         get_neighbors: Returns the objects surrounding a given cell.
     """
 
-    @staticmethod
-    def default_val():
-        """ Default value for new cell elements. """
-        return set()
-
-    def _place_agent(self, pos: Coordinate, agent: Agent) -> None:
-        """ Place the agent at the correct location. """
-        x, y = pos
-        self.grid[x][y].add(agent)
-        self.empties.discard(pos)
-
-    def _remove_agent(self, pos: Coordinate, agent: Agent) -> None:
-        """ Remove the agent from the given location. """
-        x, y = pos
-        self.grid[x][y].remove(agent)
-        if self.is_cell_empty(pos):
-            self.empties.add(pos)
-
-    @accept_tuple_argument
-    def iter_cell_list_contents(self, cell_list):
-        """
-        Args:
-            cell_list: Array-like of (x, y) tuples, or single tuple.
-
-        Returns:
-            A iterator of the contents of the cells identified in cell_list
-
-        """
-        return itertools.chain.from_iterable(
-            self[x][y] for x, y in cell_list if not self.is_cell_empty((x, y)))
+    def place_agent(self, agent: Agent, pos: Coordinate) -> Agent:
+        if not self.is_cell_empty(pos):
+            x, y = pos
+            self._grid[x][y].clear()
+            self._empties.add(pos)
+        return super().place_agent(agent, pos)
 
 
 class HexGrid(Grid):
@@ -479,8 +495,7 @@ class HexGrid(Grid):
 
     """
 
-    def iter_neighborhood(self, pos,
-                          include_center=False, radius=1):
+    def iter_neighborhood(self, pos, include_center=False, radius=1):
         """ Return an iterator over cell coordinates that are in the
         neighborhood of a certain point.
 
@@ -518,16 +533,14 @@ class HexGrid(Grid):
                 adjacent.append(pos)
 
             if x % 2 == 0:
-                adjacent += [(x - 1, y + 1), (x - 1, y),
-                             (x + 1, y + 1), (x + 1, y)]
+                adjacent += [(x - 1, y + 1), (x - 1, y), (x + 1, y + 1), (x + 1, y)]
             else:
-                adjacent += [(x - 1, y), (x - 1, y - 1),
-                             (x + 1, y), (x + 1, y - 1)]
+                adjacent += [(x - 1, y), (x - 1, y - 1), (x + 1, y), (x + 1, y - 1)]
 
             if self.torus is False:
                 adjacent = list(
-                    filter(lambda coords:
-                           not self.out_of_bounds(coords), adjacent))
+                    filter(lambda coords: not self.out_of_bounds(coords), adjacent)
+                )
             else:
                 adjacent = [torus_adj_2d(coord) for coord in adjacent]
 
@@ -554,8 +567,7 @@ class HexGrid(Grid):
         neighborhood = self.iter_neighborhood(pos)
         return self.iter_cell_list_contents(neighborhood)
 
-    def get_neighborhood(self, pos,
-                         include_center=False, radius=1):
+    def get_neighborhood(self, pos, include_center=False, radius=1):
         """ Return a list of cells that are in the neighborhood of a
         certain point.
 
@@ -572,8 +584,7 @@ class HexGrid(Grid):
         """
         return list(self.iter_neighborhood(pos, include_center, radius))
 
-    def iter_neighbors(self, pos,
-                       include_center=False, radius=1):
+    def iter_neighbors(self, pos, include_center=False, radius=1):
         """ Return an iterator over neighbors to a certain point.
 
         Args:
@@ -587,12 +598,10 @@ class HexGrid(Grid):
             An iterator of non-None objects in the given neighborhood
 
         """
-        neighborhood = self.iter_neighborhood(
-            pos, include_center, radius)
+        neighborhood = self.iter_neighborhood(pos, include_center, radius)
         return self.iter_cell_list_contents(neighborhood)
 
-    def get_neighbors(self, pos,
-                      include_center=False, radius=1):
+    def get_neighbors(self, pos, include_center=False, radius=1):
         """ Return a list of neighbors to a certain point.
 
         Args:
@@ -606,8 +615,7 @@ class HexGrid(Grid):
             A list of non-None objects in the given neighborhood
 
         """
-        return list(self.iter_neighbors(
-            pos, include_center, radius))
+        return list(self.iter_neighbors(pos, include_center, radius))
 
 
 class ContinuousSpace:
@@ -618,6 +626,7 @@ class ContinuousSpace:
     to store agent objects, to speed up neighborhood lookups.
 
     """
+
     _grid = None
 
     def __init__(self, x_max, y_max, torus, x_min=0, y_min=0):
@@ -714,8 +723,10 @@ class ContinuousSpace:
             deltas = np.minimum(deltas, self.size - deltas)
         dists = deltas[:, 0] ** 2 + deltas[:, 1] ** 2
 
-        idxs, = np.where(dists <= radius ** 2)
-        neighbors = [self._index_to_agent[x] for x in idxs if include_center or dists[x] > 0]
+        (idxs,) = np.where(dists <= radius ** 2)
+        neighbors = [
+            self._index_to_agent[x] for x in idxs if include_center or dists[x] > 0
+        ]
         return neighbors
 
     def get_heading(self, pos_1, pos_2):
@@ -777,8 +788,7 @@ class ContinuousSpace:
     def out_of_bounds(self, pos):
         """ Check if a point is out of bounds. """
         x, y = pos
-        return (x < self.x_min or x >= self.x_max or
-                y < self.y_min or y >= self.y_max)
+        return x < self.x_min or x >= self.x_max or y < self.y_min or y >= self.y_max
 
 
 class NetworkGrid:
@@ -787,7 +797,7 @@ class NetworkGrid:
     def __init__(self, G):
         self.G = G
         for node_id in self.G.nodes:
-            G.nodes[node_id]['agent'] = list()
+            G.nodes[node_id]["agent"] = list()
 
     def place_agent(self, agent, node_id):
         """ Place a agent in a node. """
@@ -814,16 +824,16 @@ class NetworkGrid:
     def _place_agent(self, agent, node_id):
         """ Place the agent at the correct node. """
 
-        self.G.nodes[node_id]['agent'].append(agent)
+        self.G.nodes[node_id]["agent"].append(agent)
 
     def _remove_agent(self, agent, node_id):
         """ Remove an agent from a node. """
 
-        self.G.nodes[node_id]['agent'].remove(agent)
+        self.G.nodes[node_id]["agent"].remove(agent)
 
     def is_cell_empty(self, node_id):
         """ Returns a bool of the contents of a cell. """
-        return not self.G.nodes[node_id]['agent']
+        return not self.G.nodes[node_id]["agent"]
 
     def get_cell_list_contents(self, cell_list):
         return list(self.iter_cell_list_contents(cell_list))
@@ -832,5 +842,9 @@ class NetworkGrid:
         return list(self.iter_cell_list_contents(self.G))
 
     def iter_cell_list_contents(self, cell_list):
-        list_of_lists = [self.G.nodes[node_id]['agent'] for node_id in cell_list if not self.is_cell_empty(node_id)]
+        list_of_lists = [
+            self.G.nodes[node_id]["agent"]
+            for node_id in cell_list
+            if not self.is_cell_empty(node_id)
+        ]
         return [item for sublist in list_of_lists for item in sublist]
